requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==12.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pillow==11.0.0
python-multipart==0.0.20
pycryptodome==3.20.0
